rename and typescript update
