+++ client/src/tools/useFetchWithAuth.ts
