+++ server/src/modules/user/userRepository.ts
@@ -1,12 +1,9 @@
 import databaseClient, { type Rows } from "../../../database/client";
 import type { Result } from "../../../database/client";
 import type { User } from "../../types/user";
-import type { UserFromDB } from "../../types/user";
 
 class userRepository {
-  async readByEmailWithPassword(
-    email: string,
-  ): Promise<UserFromDB | undefined> {
+  async findByEmail(email: string): Promise<User | undefined> {
     const [user] = await databaseClient.query<Rows>(
       "SELECT id, email, hashed_password FROM user WHERE email = ?",
       [email],
@@ -14,10 +11,10 @@ class userRepository {
 
     if (user.length === 0) return undefined;
 
-    return user[0] as UserFromDB;
+    return user[0] as User;
   }
 
-  async create(user: User) {
+  async create(user: Omit<User, "id">) {
     const [result] = await databaseClient.query<Result>(
       "INSERT INTO user (firstname, lastname, username, city, email, hashed_password) values (?, ?, ?, ?, ?, ?)",
       [
